Add function to assert that the only descriptors we have open are the ones we expect to be open. Also assert that they point at expected type.
Because openlog(3) API is unable to tell us descriptor number it is using, we
have to close syslog socket, remember assert message in local buffer and if we
fail on assertion, reopen syslog socket and log the message.

Use newly added descriptors_assert() function to ensure only expected
descriptors are open.

Close more descriptors that can be open if the worker process for the given
resource is already running.

Taken from HAST (r218045, r218044, r218370) with some modifications.